+++ scrapli/logging.py
@@ -261,6 +261,7 @@ def enable_basic_logging(
     level: str = "info",
     caller_info: bool = False,
     buffer_log: bool = True,
+    mode: str = "write",
 ) -> None:
     """
     Enable opinionated logging for scrapli
@@ -271,18 +272,26 @@ def enable_basic_logging(
         level: string name of logging level to use, i.e. "info", "debug", etc.
         caller_info: add info about module/function/line in the log entry
         buffer_log: buffer log read outputs
+        mode: string of "write" or "append"
 
     Returns:
         None
 
     Raises:
-        N/A
+        ScrapliException: if invalid mode is passed
 
     """
     logger.propagate = False
     logger.setLevel(level=level.upper())
 
     scrapli_formatter = ScrapliFormatter(caller_info=caller_info)
+
+    if mode.lower() not in (
+        "write",
+        "append",
+    ):
+        raise ScrapliException("logging file 'mode' must be 'write' or 'append'!")
+    file_mode = "a" if mode.lower() == "append" else "w"
 
     if file:
         if isinstance(file, bool):
@@ -291,9 +300,9 @@ def enable_basic_logging(
             filename = file
 
         if not buffer_log:
-            fh = FileHandler(filename=filename, mode="w")
+            fh = FileHandler(filename=filename, mode=file_mode)
         else:
-            fh = ScrapliFileHandler(filename=filename, mode="w")
+            fh = ScrapliFileHandler(filename=filename, mode=file_mode)
 
         fh.setFormatter(scrapli_formatter)